work in progress for state machine
