+++ src/main/java/frc/robot/subsystems/led/Events.java
@@ -1,0 +1,16 @@
+package frc.robot.subsystems.led;
+
+public enum Events {
+    NICK,
+    CORALINTAKE,
+    CORALSTOWED,
+    ALGAEINTAKE,
+    IDLETELEOP,
+    ELEVATORMOVING,
+    ELEVATORSCORE,
+    OFF,
+    SINGLEFADE,
+    LOS,
+    PROGRESSBAR,
+    CLEARPROGRESS
+}
